--- a/app/src/app/error.tsx
+++ b/app/src/app/error.tsx
@@ -1,0 +1,5 @@
+'use client'
+
+import ErrorBoundary from '@/components/ErrorBoundary'
+
+export default ErrorBoundary