--- a/app/src/components/providers/SessionProvider.tsx
+++ b/app/src/components/providers/SessionProvider.tsx
@@ -1,0 +1,11 @@
+'use client'
+
+interface SessionProviderProps {
+  children: React.ReactNode
+}
+
+export function SessionProvider({ children }: SessionProviderProps) {
+  // Simple provider that just wraps children
+  // All session logic is handled by the hooks
+  return <>{children}</>
+}