… .claude/settings.local.json
… .serena/cache/typescript/document_symbols_cache_v23-06-25.pkl
… API_CONSOLIDATION_COMPLETE.md
… API_CONSOLIDATION_PLAN.md
… BACKEND_CLEANUP_PLAN.md
… FRONTEND_UPDATE_GUIDE.md
… GMAIL_INTEGRATION_FIX.md
… README.md
… SETUP.md
… app/api/_lib/ai/rate-limiter.ts
… app/api/_lib/config/index.ts
… app/api/_lib/config/user.ts
… app/api/_lib/config/validate-env.ts
… app/api/_lib/db/index.ts
… app/api/_lib/db/queries/email-queries.ts
… app/api/_lib/db/queries/job-queries.ts
… app/api/_lib/db/queries/resume-queries.ts
… app/api/_lib/db/queries/thread-queries.sql
… app/api/_lib/db/types.ts
… app/api/_lib/gmail/ai-processor.ts
… app/api/_lib/gmail/conversation-analyzer.ts
… app/api/_lib/gmail/crypto.ts
… app/api/_lib/gmail/email-processor.ts
… app/api/_lib/gmail/email-taxonomy.ts
… app/api/_lib/gmail/enhanced-labeling.ts
… app/api/_lib/gmail/index.ts
… app/api/_lib/gmail/job-linker.ts
… app/api/_lib/gmail/local-token-storage.ts
… app/api/_lib/gmail/oauth.ts
… app/api/_lib/gmail/pubsub.ts